Finalizing SHACL-JS JSConstraint and JSConstraintComponent Enable functionality to map JSConstraintComponent parameters into the sh:message string Pull out JSExecutable into its own file Pull out JSConstraintComponent and related classes into their own file Fix bug in SPARQLConstraintComponent parameters mapping into sh:message
