+++ pyshacl/constraints/sparql/sparql_based_constraint_components.py
@@ -60,6 +60,9 @@ class BoundShapeValidatorComponent(ConstraintComponent):
         # TODO:coverage: this is never used for this constraint?
         return SH_ConstraintComponent
 
+    def make_generic_messages(self, datagraph: GraphLike, focus_node, value_node) -> List[rdflib.Literal]:
+        return [rdflib.Literal("Parameterised SHACL Query generated constraint validation reports.")]
+
     def evaluate(self, target_graph: GraphLike, focus_value_nodes: Dict, _evaluation_path: List):
         """
         :type focus_value_nodes: dict
@@ -67,7 +70,7 @@ class BoundShapeValidatorComponent(ConstraintComponent):
         """
         reports = []
         non_conformant = False
-        extra_messages = self.query_helper.messages or None
+        extra_messages = self.constraint.messages or []
         rept_kwargs = {
             # TODO, determine if we need sourceConstraint here
             #  'source_constraint': self.validator.node,
@@ -81,22 +84,43 @@ class BoundShapeValidatorComponent(ConstraintComponent):
                 violations = self.validator.validate(f, value_nodes, target_graph, self.query_helper)
             except ValidationFailure as e:
                 raise e
-            if not self.shape.is_property_shape:
-                result_val = f
-            else:
-                result_val = None
-            for v in violations:
+            for val, vio in violations:
                 non_conformant = True
-                if isinstance(v, bool) and v is True:
-                    # TODO:coverage: No test for when violation is `True`
-                    rept = self.make_v_result(target_graph, f, value_node=result_val, **rept_kwargs)
-                elif isinstance(v, tuple):
-                    t, p, v = v
+                msg_args_map = self.query_helper.param_bind_map.copy()
+                msg_args_map.update({"this": f, "value": val})
+                if self.shape.is_property_shape:
+                    msg_args_map['path'] = self.shape.path()
+                self.query_helper.bind_messages(msg_args_map)
+                bound_messages = self.query_helper.bound_messages
+                # The DASH test suite likes _no_ value entry in the report if we're on a Property Shape.
+                report_val = val if not self.shape.is_property_shape else None
+                if isinstance(vio, bool):
+                    if vio is False:  # ASKValidator Result
+                        new_kwargs = rept_kwargs.copy()
+                        new_kwargs['extra_messages'].extend(bound_messages)
+                        rept = self.make_v_result(target_graph, f, value_node=report_val, **new_kwargs)
+                    else:  # SELECTValidator Failure
+                        raise ValidationFailure("Validation Failure generated by SPARQLConstraint.")
+                elif isinstance(vio, tuple):
+                    t, p, v = vio
+                    new_msg_args_map = msg_args_map.copy()
                     if v is None:
-                        v = result_val
-                    rept = self.make_v_result(target_graph, t or f, value_node=v, result_path=p, **rept_kwargs)
+                        v = report_val
+                    else:
+                        new_msg_args_map['value'] = v
+                    if p is not None:
+                        new_msg_args_map['path'] = p
+                    if t is not None:
+                        new_msg_args_map['this'] = t
+                    self.query_helper.bind_messages(new_msg_args_map)
+                    new_bound_msgs = self.query_helper.bound_messages
+                    new_kwargs = rept_kwargs.copy()
+                    new_kwargs['extra_messages'].extend(new_bound_msgs)
+                    rept = self.make_v_result(target_graph, t or f, value_node=v, result_path=p, **new_kwargs)
                 else:
-                    rept = self.make_v_result(target_graph, f, value_node=v, **rept_kwargs)
+                    new_kwargs = rept_kwargs.copy()
+                    new_kwargs['extra_messages'].extend(bound_messages)
+                    rept = self.make_v_result(target_graph, f, value_node=report_val, **new_kwargs)
                 reports.append(rept)
         return (not non_conformant), reports
 
@@ -182,6 +206,20 @@ class SPARQLConstraintComponentValidator(object):
         self.messages = message_nodes
         self.initialised = True
 
+    def make_messages(self, params_map=None):
+        if params_map is None:
+            return self.messages
+        ret_msgs = []
+        for m in self.messages:
+            this_m = m.value[:]
+            for a, v in params_map.items():
+                replace_me = "{$" + str(a) + "}"
+                if isinstance(v, rdflib.Literal):
+                    v = v.value
+                this_m = this_m.replace(replace_me, str(v))
+            ret_msgs.append(rdflib.Literal(this_m))
+        return ret_msgs
+
 
 class AskConstraintValidator(SPARQLConstraintComponentValidator):
     def __new__(cls, shacl_graph: 'ShapesGraph', node, *args, **kwargs):
@@ -221,7 +259,7 @@ class AskConstraintValidator(SPARQLConstraintComponentValidator):
         new_bind_vals = new_bind_vals or {}
         bind_vals = param_bind_vals.copy()
         bind_vals.update(new_bind_vals)
-        violations = []
+        violations = set()
         for v in value_nodes:
             if query_helper is None:
                 # TODO:coverage: No test for this case when query_helper is None
@@ -240,7 +278,7 @@ class AskConstraintValidator(SPARQLConstraintComponentValidator):
                 # TODO:coverage: Can this ever actually happen?
                 raise ValidationFailure("ASK Query did not return an askAnswer.")
             if answer is False:
-                violations.append(v)
+                violations.add((v, False))
         return violations
 
 
@@ -303,23 +341,24 @@ class SelectConstraintValidator(SPARQLConstraintComponentValidator):
                 except KeyError:
                     p = None
                 try:
-                    v = r['value']
+                    v2 = r['value']
                 except KeyError:
-                    v = None
+                    v2 = None
                 try:
                     t = r['this']
                 except KeyError:
                     # TODO:coverage: No test for when result has no 'this' key
                     t = None
-                if p or v or t:
-                    violations.add((t, p, v))
+                if p or v2 or t:
+                    violations.add((v, (t, p, v2)))
                 else:
                     # TODO:coverage: No test for generic failure, when
                     #  'path' and 'value' and 'this' are not returned.
                     #  here 'failure' must exist
                     try:
-                        _ = r['failure']
-                        violations.add(True)
+                        f = r['failure']
+                        if f is True or (isinstance(f, rdflib.Literal) and f.value):
+                            violations.add((v, True))
                     except KeyError:
                         pass
         return violations
@@ -337,6 +376,11 @@ class SPARQLConstraintComponent(CustomConstraintComponent):
         return super(SPARQLConstraintComponent, cls).__new__(
             cls, shacl_graph, node, parameters, validators, node_validators, property_validators
         )
+
+    @property
+    def messages(self):
+        # TODO: allow messages at this SPARQLConstraintComponent level
+        return []
 
     def make_validator_for_shape(self, shape: 'Shape'):
         """
